+++ lib/support_table_data/documentation.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module SupportTableData
+  module Documentation
+  end
+end
+
+require_relative "documentation/source_file"
+require_relative "documentation/yard_doc"