+++ spec/models/alias.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class Alias < ActiveRecord::Base
+  belongs_to :color
+
+  validates_uniqueness_of :name
+end