+++ spec/models/group.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+class Group < ActiveRecord::Base
+  include SupportTableData
+
+  self.primary_key = :group_id
+
+  named_instance_attribute_helpers :group_id
+
+  add_support_table_data "groups.yml"
+
+  named_instance_attribute_helpers :name
+
+  validates_uniqueness_of :name
+end