+++ spec/models/shade_hue.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+class ShadeHue < ActiveRecord::Base
+  belongs_to :shade
+  belongs_to :hue
+end