+++ spec/models/triangle.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Triangle < Polygon
+  validates :side_count, numericality: {equal_to: 3}
+end